src/um/persona/Persona.java
package um.persona;

public class Persona {
    private String nombre;
    private int edad;
    private Persona madre;
    private Persona padre;

    public Persona(String nombre, int edad) {
        this.nombre = nombre;
        this.edad = edad;
    }

    public Persona getMadre() {
        return madre;
    }

    public void setMadre(Persona madre) {
        this.madre = madre;
    }

    public Persona getPadre() {
        return padre;
    }

    public void setPadre(Persona padre) {
        this.padre = padre;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public int getEdad() {
        return edad;
    }

    public void setEdad(int edad) {
        this.edad = edad;
    }

    public boolean mayorDeEdad() {
        boolean resultado = false;
        if (edad > 18) {
            resultado = true;
        }
        return resultado;
    }

}

